backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
python-multipart==0.0.6

# ML and AI - Basic versions
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
datasets>=2.0.0
accelerate>=0.20.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Data Science - Basic versions
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional for GPU
# torchtext==0.16.2  # if needed